refactor: dateHelper to dateUtils
